fixup! lx_fs: enable watch
This patch fixes errors that only occurred when a lot of changes were
made to a file from the Linux side in a short time.
To achieve this we now watch the parent directories of the desired
watches. By this the watch handle is still valid when an entry is
removed and later re added (e.g. `cp`).

Issue #4070