WIP: sandbox: enable CPU/PD fault detection for children
It is now possible to configure to monitor a child for
either CPU and/or PD (segmentation) faults.

This monitoring can be enabled independently of each other with
the config sub nodes `<monitor_cpu_faults/>` and `<monitor_pd_faults/>`.

If a fault is detected in a child, this is reported with one of
the attributes `cpu_fault` or `pd_fault` depending on the detected
fault.

TODO: config changes of a child not yet handles the two new sub nodes.