sandbox: enable PD fault detection for children
Start nodes now support the sub node `<monitor_pd_faults/>`.
If this node is present, a fault handler is attached to the PD-session
of the child.
If a fault in the PD of a child is detected, and state reporting is
enabled, the attribute `pd_fault` is set to true.

Issue gapfruit#1446